move all chat components to chat
